Add transitive closure table for problems. Amend pattern files.
Created a new file 'transitive.csv' which contains:
- ancestorId
- descendantId
for all ancestor / descendant pairs in the CDB. This is intended
for 'relaxed' assessment of algorithm performance - i.e. allowing
the most precise term or a less precise (ancestor) term.

The pattern files have been amended - the human generated example
files are no longer included in the patterns file and should not be
used for training as they may introduce bias, because they may cause
the algorithm to learn spurious associations.

However they are useful for testing and should be used for this purpose.